src/main/java/com/gustavo/comicreviewapi/services/AuthorizationService.java
package com.gustavo.comicreviewapi.services;

import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import com.gustavo.comicreviewapi.repositories.UserRepository;
import com.gustavo.comicreviewapi.entities.User;
import com.gustavo.comicreviewapi.utils.UserSS;

@Service
public class AuthorizationService implements UserDetailsService {
	
	private UserRepository userRepository;
	
	public AuthorizationService(UserRepository userRepository) {
		this.userRepository = userRepository;
	}
	
	@Override
	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
		User user = userRepository.findByEmail(username);
		
		if (user == null) {
			throw new UsernameNotFoundException(username);
		}
		
		return new UserSS(user.getId(), user.getEmail(), user.getPassword(), user.getProfiles());
	}

}
